API: update unit rate to decimal. Mapping fixes/updates

+++ drr/src/API/EMCR.DRR/Managers/Intake/Contract.cs
@@ -631,7 +631,7 @@ namespace EMCR.DRR.Managers.Intake
         public ResourceCategory? Resources { get; set; }
         public CostUnit? Units { get; set; }
         public decimal? Quantity { get; set; }
-        public int? UnitRate { get; set; }
+        public decimal? UnitRate { get; set; }
         public decimal? TotalCost { get; set; }
         public int? TaskNumber { get; set; }
     }
